fix bug with large IDs and some simple routes for CRUD offers

+++ database/migrations/2016_04_23_020434_create_offers_table.php
@@ -14,7 +14,7 @@ class CreateOffersTable extends Migration
     {
         Schema::create('offers', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('f_b_user_id')->unsigned();
+            $table->bigInteger('f_b_user_id')->unsigned();
             $table->foreign('f_b_user_id')->references('id')->on('f_b_users');
             $table->string('food');
             $table->dateTime('time');
